feat(input and markup): Runtime command validator
This change adds a new element to the language of specifying lessons.
Any expected value that is prefixed by the '!' char will be interpreted
as a command and validated at run time. This neccessitates an adaptation
to the input filter to check the users command vs the expected command
at run time.

This ensures for example when the user performs ls it is being validated
against the cwd of the user at runtime.