feat(tui): Address scrollback issue and implement mouse in menu
Alt screen buffer causes an issue due to scroll buffer implementation
being inconsistent across terminals when it comes to the alt buffer.

This now sets the application to not launch in the alt screen buffer and
instead elects to "fake it" using some screen clearing. Mouse messages
from bubble tea must be disabled in the lesson menu as they just print
out escape sequences.

Mouse use is also now available in the menu view. One known limitation
of implementing a double-click feature is that click dragging does not
work.